Implemented alternative measurement selector for the CKF.
This measurement selector assumes that measurements associated to
one surface all have the same dimension and have the same mapping
to the bound parameters. Moreover it is assumed that the number of
measurements per surface is large (>~10) but only few measurements
O(1) are compatible with the current trajectory.

The selector will first gather N calibrated measurements with the
smallest chi2, and only create track states for the finally selected
measurements. Also since the mapping between the bound parameters and
the measurement is the same for all measurements associated to a
surface the prediction in the measurement domain is only computed
once per surface and trajectory.